Implemented the Spike Epic 1 on a branch not multiple commits behind main.

+++ base/views.py
@@ -1,6 +1,6 @@
 from multiprocessing import context
 
-from django.http import HttpResponseRedirect
+from django.contrib.auth.decorators import login_required
 from django.shortcuts import render, redirect
 from django.contrib import messages
 from django.contrib.auth.models import User
@@ -63,75 +63,96 @@ def logoutUser(request):
     return redirect('home')
 
 
+@login_required(login_url='/login')
 def home(request):
     rooms = Room.objects.all()
     sheets = Sheet.objects.all()
     context = {'rooms': rooms, 'sheets': sheets}
     return render(request, 'base/home.html', context)
 
+
+@login_required(login_url='/login')
 def glossary(request):
     room = Room.objects.all()
     context = {'room': room}
     return render(request, 'base/glossary.html', context)
 
 
+@login_required(login_url='/login')
 def room(request, pk):
     room = Room.objects.get(id=pk)
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
 
+@login_required(login_url='/login')
 def sheet(request, pk):
     sheet = Sheet.objects.get(id=pk)
+
+    if request.user != sheet.user and not request.user.is_superuser:  # only admins or users can view their own sheets
+        return redirect('home')
+
     context = {'sheet': sheet}
     return render(request, 'base/sheet.html', context)
 
 
-def createSheet(request):
+@login_required(login_url='/login')  # can only be used by users that are logged in, else redirect to login page
+def create_sheet(request):
     """
     Create sheet and save if valid.
     :param request: object to pass state to the function
-    :return: url to return to
+    :return: url to redirect the user to
     """
     form = SheetForm()  # get default form
+
     if request.method == 'POST':
         form = SheetForm(request.POST)  # get form for post
         if form.is_valid():  # if valid, save, then redirect
+            form = form.save(commit=False)  # save the form, then update user to user who created the sheet
+            form.user = request.user
             form.save()
-            return HttpResponseRedirect('home')
+            return redirect('home')
 
     context = {'form': form}
     return render(request, 'base/sheet_form.html', context)
 
 
-def updateSheet(request, pk):
+@login_required(login_url='/login')  # can only be used by users that are logged in, else redirect to login page
+def update_sheet(request, pk):
     """
     Edit the sheet and saves the data if valid.
     :param request: object to pass state to the function
     :param pk: ID for the sheet
-    :return: url to return to
+    :return: url to redirect the user to
     """
     sheet = Sheet.objects.get(id=pk)  # get sheet with ID
     form = SheetForm(instance=sheet)  # get current form
 
+    if request.user != sheet.user and not request.user.is_superuser:  # only admins or users can edit their own sheets
+        return redirect('home')
+
     if request.method == 'POST':
-        origin = request.POST.get('origin', '/')  # get prev location to redirect
         form = SheetForm(request.POST, instance=sheet)  # get form for post
         if form.is_valid():  # if valid, save, then redirect to previous location
             form.save()
             return redirect(f'/sheet/{pk}')
 
     context = {'form': form}
-    return render(request, 'base/sheet_form.html', context)  # otherwise don't leave
+    return render(request, 'base/sheet_form.html', context)  # render existing form
 
 
-def deleteSheet(request, pk):
+@login_required(login_url='/login')  # can only be used by users that are logged in, else redirect to login page
+def delete_sheet(request, pk):
     """
     Delete sheet on user confirmation.
     :param request: object to pass state to the function
-    :return: url to return to
+    :return: url to redirect the user to
     """
     sheet = Sheet.objects.get(id=pk)
+
+    if request.user != sheet.user and not request.user.is_superuser:  # only admins or users can delete their own sheets
+        return redirect('home')
+
     if request.method == 'POST':
         sheet.delete()
         return redirect('home')
